personal code review changes
