+++ src/main/java/edu/cornell/kfs/concur/batch/service/ConcurStandardAccountingExtractValidationService.java
@@ -15,10 +15,9 @@ public interface ConcurStandardAccountingExtractValidationService {
     
     boolean validateDebitCreditField(String debitCredit);
     
-    boolean validateEmployeeGroupId(String employeeGroupId);
-    
     boolean validateDate(Date date);
     
     boolean validateConcurStandardAccountingExtractDetailLine(ConcurStandardAccountingExtractDetailLine line);
-
+    
+    boolean validateEmployeeGroupId(String employeeGroupId);
 }
